Fixed auth api, started etcd from fixture, separated tests
