+++ jupyterhub_traefik_proxy/traefik_utils.py
@@ -7,7 +7,7 @@ from urllib.parse import urlparse
 from tornado.httpclient import AsyncHTTPClient, HTTPRequest
 
 
-async def check_traefik_dynamic_conf_ready(api_url, target):
+async def check_traefik_dynamic_conf_ready(username, psw, api_url, target):
     """ Check if traefik loaded its dynamic configuration from the
         etcd cluster """
     expected_backend = create_backend_alias_from_url(target)
@@ -15,10 +15,14 @@ async def check_traefik_dynamic_conf_ready(api_url, target):
     ready = False
     try:
         resp_backends = await AsyncHTTPClient().fetch(
-            api_url + "/api/providers/etcdv3/backends"
+            api_url + "/api/providers/etcdv3/backends",
+            auth_username=username,
+            auth_password=psw,
         )
         resp_frontends = await AsyncHTTPClient().fetch(
-            api_url + "/api/providers/etcdv3/frontends"
+            api_url + "/api/providers/etcdv3/frontends",
+            auth_username=username,
+            auth_password=psw,
         )
         backends_data = json.loads(resp_backends.body)
         frontends_data = json.loads(resp_frontends.body)
@@ -35,11 +39,13 @@ async def check_traefik_dynamic_conf_ready(api_url, target):
         return ready
 
 
-async def check_traefik_static_conf_ready(api_url):
+async def check_traefik_static_conf_ready(username, psw, api_url):
     """ Check if traefik loaded its static configuration from the
     etcd cluster """
     try:
-        resp = await AsyncHTTPClient().fetch(api_url + "/api/providers/etcdv3")
+        resp = await AsyncHTTPClient().fetch(
+            api_url + "/api/providers/etcdv3", auth_username=username, auth_password=psw
+        )
         rc = resp.code
     except ConnectionRefusedError:
         rc = None
